Implement hot reloading
This commit restructures the docker files, there is now a separate dev
and production dockerfile/compose-file. Hotreloading is done with
codedaemon which has been added as a dependency.
The project folder has been added as a volume in order to detect file changes.
This means that the compiled binaries will show up on the host machine,
but as these should not be committed, they have been moved to a
gitignored directory bin.