Implement shell script that initilizes the cluster
init-cluster.sh initilizes all the containers in the swarm by
initilizing each containers kademlia node. This is done by inspecting
each container and getting their IPs and then executing the node init
method to setup the node.

Finally, every node except one gets a known kademlia node added to their
routing table as a contact. This allows for the nodes to join the
network once we have implemented the FIND_NODE RPC later.

To implement this three new commands were added:
- getid: The getid command returns the node ID of the node running inside the container

- addcontact: Added a new command which adds a new contact to the nodes routing table given the nodeID and address of the other node.

- init: init takes the nodes local address as an argument since this is
required for the node to create a new routing table with itself as a
contact.