Implement network joining and FIND_NODE RPC
The lookup algorithm was implemented by following the description in the
kademlia specification. Note that the current version is still missing
some key features of the lookup algorithm since we are using alpha=1
which makes it non-parallel. This will be fixed once we have merged
Martins FIND_VALUE branch as well since we need to fix some underlying
issues to get the algortihm working with alpha>1. The logic behind what
happens if no improvement of the closest node has been made after one
iteration also has to be implemented later.

Added CLI commands:
- Join: For joining the network, only executes a node lookup on itself
to fill the nodes routing table with entries.

Added RPCs:
- findnode: Used by the node lookup algorithm to search on the specified
key (id) this is the FIND_NODE RPC.
- findnoderesp: The response of a findnode request. This response
contains k contacts in its content with format <ID>!<Address>