Implement get new ID in range of specified bucket
Implemented funcationality for generating a new NodeID which will be in
the range of a nodes bucket *i*. The implementation is a bit messy and
hard to understand but what I basically do is loop through each byte in
the byte array of the nodes id (kademliaID is represented as a byte
array). And for each byte I copy the bits. This
continues until we reach bit BucketIndex+1 where we make sure that the
bit differs (otherwise the new ID would not be placed in the specified
bucket) then we generate random bits for the remaining bits of the new
ID.

Remember that a Contact will be placed in bucket i if the distance from
the node to the contact is within the range 2^i <= Distance(node,
contact) < 2^i+1.

See my comments for why I have to reverse the bucketIndex.