Add bucket refresh step to network joining
The node now refreshes every bucket further away than the bucket its
closest neighbour resides in. This step is done after the initial node
lookup has been completed. The closest neighbour node is the first
contact in the list of contacts returned from the node lookup.

A bucket refresh is done by generating a new random ID within the
range of the bucket (for bucket i this is 2^i <= Distance(node, contact) < 2^(i+1))
and then performing a node lookup on that ID. This is repeated for all
buckets.

This commit also fixes an issue I noticed while testing this where we
don't create enough channels for the final round of FIND_* RPCs if the
closest contact did not change during the iteration. Fixed it by
creating k channels instead of alpha.