Implement method WithLock on rpc pool
This is simply a higher order function that takes a function f and locks
the rpc pool's lock before calling f, and then unlocks the lock after f
has finished executing. This means the lock doesn't manually have to be
locked/unlocked and thus eliminates the risk of forgetting to
lock/unlock.