Implement TTL delay
Now, when inserting into the datastore you have to pass a contact
pointer. If this pointer is non-nil then the node that is inserting is the node
where the insert was initiated, and if it is nil then it is not. This
means that this pointer will be non-nil when a node receives STORE
rpc-command and nil when a node receives the put cli-command.

Introduces a new rpc called REFRESH, which contains the hash of the file
to refresh and is used to tell another node to restart its refresh timer.
Also, each datastore entry takes a list of contacts that indicate at
which other contacts the value was stored when initially inserted. This
is used to remember which nodes to send REFRESH rpcs to.