Implement forget cli-command
Each entry in the datastore now has a bool that indicates whether or not
the value has been marked as forgotten. If a value has been marked as
forgotten, then it will be deleted and no REFRESH rpcs will be sent to
the other node during the next refresh timer interval and the timer will
be stopped.