Add tests for main functions in the lookup algo
Added some tests to cover the branches of the lookup algo. This was done
by making the SetupLookUpAlgorithm function mockable. We can now mock it
and return variables which we can use during the test. This way we are
able to send data to the channels the lookup algo is waiting for to
simulate responses from other nodes. This is not the best approach but
at least we can test it now even though it requires a lot of waits and
is very tedious and error prone.